improved parameterization, removed pins
Added constants, to parameterize the scanner diffently.
Amongs others, the SCAN_LINE is now determined from
when the exposure starts and stops. It also depends on the
rotations of the polygon per minute.
Laserpin and polypon no longer used but controlled directly from
pru. They have been removed. Still need to test code on the Hexastorm.