improvement stabilizer
if the laser head did not get data it would get locked in an await for data loop
this disabled the possibility of multiple exposures.
It would simply not pass the sync part. If no data is sent to the laser the behavior is still strange.
It keeps spinning but does not sync --> check this out